+++ src/main/java/hermione/commands/DeadlineCommand.java
@@ -6,8 +6,8 @@ import hermione.exceptions.TaskValidationException;
 import hermione.storage.TaskStorage;
 import hermione.tasks.Deadline;
 import hermione.tasks.Task;
-import hermione.ui.common.UiUtils;
 import hermione.utils.DateUtils;
+import hermione.utils.UiUtils;
 
 /**
  * Represents a command to create a new Deadline task in the Hermione application.
