+++ src/main/java/hermione/ui/javafx/Main.java
@@ -1,4 +1,4 @@
-package hermione.ui.fx;
+package hermione.ui.javafx;
 
 import java.io.IOException;
