+++ src/main/resources/view/MainWindow.fxml
@@ -10,7 +10,7 @@
 <AnchorPane xmlns:fx="http://javafx.com/fxml/1" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity"
             minWidth="-Infinity"
             prefHeight="600.0" prefWidth="400.0" stylesheets="@../css/main.css"
-            xmlns="http://javafx.com/javafx/24.0.1" fx:controller="hermione.ui.fx.MainWindow">
+            xmlns="http://javafx.com/javafx/24.0.1" fx:controller="hermione.ui.javafx.MainWindow">
     <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
                AnchorPane.bottomAnchor="1.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="76.0"/>
     <Button fx:id="sendButton" layoutX="324.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput"
